// Users/darragh/Documents/gmp/firebase-studio-prototyper-test-2/test-project/test-project/src/lib/types.ts
export interface MappedShopifyProduct {
  id: string;
  title: string;
  vendor: string;
  product_type: string;
  price: number;
  inventory: number;
  imageUrl: string;
  imageHint: string;
}